render prop used for treatments
